requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.1
bs4==0.0.1
dj-database-url==0.5.0
Django==2.1
django-crontab==0.7.1
django-heroku==0.3.1
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
pytz==2020.1
soupsieve==2.0.1
sqlparse==0.3.1
whitenoise==5.1.0
